+++ datasette/app.py
@@ -1,6 +1,4 @@
 import asyncio
-import aiofiles
-from mimetypes import guess_type
 import collections
 import hashlib
 import json
@@ -38,6 +36,7 @@ from .utils import (
     sqlite_timelimit,
     to_css_class,
 )
+from .utils.asgi import asgi_static, asgi_send_html, asgi_send_json
 from .tracer import capture_traces, trace
 from .plugins import pm, DEFAULT_PLUGINS
 from .version import __version__
@@ -714,78 +713,3 @@ class Datasette:
         #             await database.table_counts(limit=60 * 60 * 1000)
 
         return app
-
-
-async def asgi_send_json(send, info, status=200, headers=None):
-    headers = headers or {}
-    await asgi_send(
-        send,
-        json.dumps(info),
-        status=status,
-        headers=headers,
-        content_type="application/json",
-    )
-
-
-async def asgi_send_html(send, html, status=200, headers=None):
-    headers = headers or {}
-    await asgi_send(
-        send, html, status=status, headers=headers, content_type="text/html"
-    )
-
-
-async def asgi_send(send, content, status, headers, content_type="text/plain"):
-    await asgi_start(send, status, headers, content_type)
-    await send({"type": "http.response.body", "body": content.encode("utf8")})
-
-
-async def asgi_start(send, status, headers, content_type="text/plain"):
-    # Remove any existing content-type header
-    headers = dict([(k, v) for k, v in headers.items() if k.lower() != "content-type"])
-    headers["content-type"] = content_type
-    await send(
-        {
-            "type": "http.response.start",
-            "status": status,
-            "headers": [
-                [key.encode("latin1"), value.encode("latin1")]
-                for key, value in headers.items()
-            ],
-        }
-    )
-
-
-def asgi_static(root_path, chunk_size=4096):
-    async def inner_static(scope, receive, send):
-        path = scope["url_route"]["kwargs"]["path"]
-        full_path = (Path(root_path) / path).absolute()
-        # Ensure full_path is within root_path to avoid weird "../" tricks
-        try:
-            full_path.relative_to(root_path)
-        except ValueError:
-            await asgi_send_html(send, "404", 404)
-            return
-        first = True
-        try:
-            async with aiofiles.open(full_path, mode="rb") as fp:
-                if first:
-                    await asgi_start(
-                        send, 200, {}, guess_type(str(full_path))[0] or "text/plain"
-                    )
-                    first = False
-                more_body = True
-                while more_body:
-                    chunk = await fp.read(chunk_size)
-                    more_body = len(chunk) == chunk_size
-                    await send(
-                        {
-                            "type": "http.response.body",
-                            "body": chunk,
-                            "more_body": more_body,
-                        }
-                    )
-        except FileNotFoundError:
-            await asgi_send_html(send, "404", 404)
-            return
-
-    return inner_static
